Fix cppcheck suppressions in LoadILLSALSA and LoadILLLagrange (#39455)
With ongoing Nexus refactor work, CppCheck noticed these lines in these
algorithms and complained about an out-of-bounds issue. This is
apparently because it doesn't recognize that `dimsSize().size()` should
be at least as big as `nDims`.

Hopefully the added check will convince CppCheck

Refs #38332 

#### Further detail of work

Adds the check
``` cpp
if (dimsSize.size() < nDims)
    throw std::runtime_error("Scanned variables are not formatted properly. Check you nexus file.");
```