Change to use Instrument 2.0
Instead of looping through the components in the instrument, we now use
the information on the DetectorInfo class on the workspace to get IDs
and position, etc. This will hopefully make things simpler and
easier to maintain in the future. This also eases the difficulties
working out how to map components to spectra, since it's easy with
detector IDs.