Enable use of a unit test thread limit in Jenkins. (#39481)
Add a check for a new `UNIT_TEST_THREADS` environment variable that can
be set in the Jenkins configuration for each node that can be used to
limit the number of unit tests to be run in parallel. This has been done
so that we can limit the number of unit tests processes on out macOS
builders, some of which have limited RAM (compared to the core count).

### Further detail of work
I have set `UNIT_TEST_THREADS=6` on the following nodes, which have been
observed to timeout on tests:
- isis-ndw2564
- isis-ndw2995
- isis-ndw2996

This value is two below the number of cores, so will hopefully give the
machines more resources to run the tests in parallel successfully. We
can tweak the number inside the Jenkins configuration if required.