Refactor Abins, replace SData with Euphonic data structures (#37350)
### Description of work

#### Summary of work
Remove `abins.sdata.SData`, use classes from `euphonic.spectra` instead

#### Purpose of work
Abins currently manages collections of simulated spectra with an "SData"
class. This assumes that the data will be structured in a tree with
heirarchy atom -> quantum-order. This is not flexible enough to deal
with coherent scattering, which might not be decomposable by atom.

Euphonic is an existing dependency and uses Spectrum1D, Spectrum2D and
Spectrum1DCollection classes to manage spectra including arbitrary
metadata. Effectively this is a flat table structure, amenable to
"query" or "pipeline" workflows. This seems to be an appropriate tool,
so here we extend that to add the missing Spectrum2DCollection and work
towards using those natively in Abins.

#### Further detail of work
- [x] Add a method to SData to produce an equivalent
SpectrumNDCollection object
- [x] Update abins to use these objects for all "output" management
(i.e. summing relevant subsets and creating workspaces)
- [x] Update abins to simulate data directly into a SpectrumNDCollection
object
- [x] Delete SData class and related unit tests
- [x] Move patches and extensions to Euphonic classes upstream into a
new euphonic release
- [x] Remove Euphonic subclasses from mantid, update Euphonic dependency
version