Crosshair in tiled plots (#39491)
This PR enables the crosshair option in 2D tiled plots and color contour
plots. The crosshair is still disabled in 3D plots.

[EWM11427](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20(Change%20Management)#action=com.ibm.team.workitem.viewWorkItem&id=11427)

If there are multiple subplots, the crosshair's movement will sync up
based on the actual data location. Zooming in and out of the figure will
allow more accurate movement of the crosshair. See details in tests.