Minor refactors for char data in napi5 (#39483)
This refactors the handling of char/string data in napi5 to be easier to
read and work more efficiently with the HDF5 C library methods. This is
meant to make future work of integrating into Nexus File with H5Cpp
easier.

The following:
- removed on the many `getAttr()` methods
- changed `getStrAttr()` to be called with the attribute's *name* and
not its `Info` object
- all the fixes to make the above ^ changes work
- cleaned up `NX5getdata()` to have fewer labyrinthine if/else blocks,
and cleanly handle char data
- more tests in NexusFileTest which safeguard several ways reading data
can go wrong when string data dimensions are inconsistent with the
actual data

Refs Issue #38332 
[EWM
11399](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20%28Change%20Management%29#action=com.ibm.team.workitem.viewWorkItem&id=11399)

#### Further detail of work

There is a single new function for reading char data,
`NX5getchardata()`. Char data is either variable or fixed-width string.

If variable string, use the hdf5 functionality that sets the proper
string array size for us, together with
[`H5free_memory()`](https://support.hdfgroup.org/documentation/hdf5/latest/group___h5.html#ga71872bf6445cba956da86d4762b662cf).

If fixed-width, use a combination of the dataspace's `dims` aray,
together with the datatypes size property, to create a proper char block
array and read data into that. There are several precautions around
incorrect or imprecise usage, which are protected by the new tests.