Refactor to use arrays instead of Detector class
This refactor does not use a Detector class but instead keeps track of
detector properties in several arrays with the same ordering and the
same size.

The main advantage of this design is that it avoids creating multiple
maps between detector IDs and other properties and allows for easily
selecting non-monitor or monitor detectors by using the boolean mask:

eg. self._detector_ids[~self._is_monitor] selects the non-monitor
detector ids

There are some small (probably negligible) performance improvements, due
to the fact that all required arrays are calculated at the beggining and
there is less need to constantly iterate over a list[Detector] to
retrieve lists of properties.
Since previously a lot of the getter methods in the model were already
retrieving lists, I think it makes more sense for the model to keep
track of these lists/arrays instead of keeping track of many different
Detector() objects.

Finally, Mantid's design of DetectorInfo is the same: DetectorInfo acts
as a big map where we can retrieve properties of detectors through
indexing. I think it might be awkward to use DetectorInfo inside a
Detector class, because technically DetectorInfo contains information on
all detectors, so I am not sure it should be contained whithin each
detector.