Handle detectors with no corresponing spectrum
On some instruments there are detectors listed in the DectectorInfo
class that do not have a corresponding spectrum in the workspace. Work
around this by using a different method to map detector IDs to
workspace indices. It already existed, I just had to add a Python
export for it. Also added a test for this situation.