AlignAndFocusPowderSlim MinMax changes (#39526)
This attempts to make small performance improvements on #39494 by using
more references and sharing data between `MinMax` structs on splitting.

Rather than reading the detid and time-of-flight one after another (disk
limited), this loads the detid, parses it for some information, then
loads the time-of-flight. This slightly helps disk access in terms of
concurrency.

There is also a small amount of code reshuffling to put constructors
next to each other and instance variables next to each other.