`Nexus` enum classes :lock: (#39585)
There are several collections of variables inside `NexsFile_fwd`, which
were defined as part of the `napi` package, which really should have
been `enum class`.

This makes them `enum class`, removes the `#define` and `constexpr`
declarations of the values a separate integers (so they are *only*
inside the `enum class`), and changes things to use the `enum` version.

Refs #38332 
EWM 11399

#### Further detail of work

It looks like a lot of changes, but it isn't.

The biggest change was to the file access mode, `NXaccess`, which was
used in about 50 files. The changes made going from `NXACC_CREATE5` -->
`NXaccess:CREATE5` was purely mechanical.

Because of the large number of files touched by this, it makes more
sense to do this standalone.

I added two methods for printing out `NXaccess` and `NXcompression`,
which will now be strings instead of integers.

I also deleted the `napi` method `NXIprintlink`. This did nothing more
than print info about a `NXlink` object, and was going to be deleted
soon anyway. It was used in one test for error results, and that test
will also be deleted soon anyway.

### To test:

This is a pure refactor, existing unit tests should be enough.

*This does not require release notes* because this is a purely internal
issue

---------

Co-authored-by: Carson Sears <searscr@ornl.gov>