Optimize the addTies function by implementing a single sort operation after all ties are added (#39293)
This PR optimises the `addTies` function in the interface of all fitting
functions `IFunction`. Previously, the function sorted the ties each
time a new tie was added. This PR changes the behaviour by sorting all
ties after they're completely added, and if sorting fails, it reverses
all changes. This implementation has created a significant performance
boost, as shown in the graph below comparing old and new
implementations. I know there's duplicate code in the implementation,
but when I tried to refactor the function, it became more complicated,
so I went back to this simpler version with minimal changes. Since this
is core code used throughout the codebase, keeping changes small is
safer. I've also added a unit test that covers all cases.