Add diagnostics checkbox to refl gui (#39587)
### Description of work

This PR adds a checkbox to the Reflectometry GUI that exposes the
ability to output intermediate workspaces for diagnostics.

### To test:

Ensure access to data archive.

1) Open Reflectometry GUI
2) On `Reduction Preview` tab, enter run `84884` then click `Load` (You
will see unrelated, unconcerning error with this dataset)
3) On the instrument view graphic, use the region selector above to
select an interesting portion of the detectors.
4) On the sliceviewer plot next to it, use the region selector above
that to select a signal range (containing the specular)
5) Using the drop own next to the signal region selector, select a
background range on the same plot.
6) Click apply, these settings will be sent to the relevant tabs.
7) On the `Experiment Settings` tab check the `Diagnostics` checkbox,
also, select the `Subtract Background` checkbox.
8) Finally, return to the runs tab. Enter a random group name in the
first row, then in the second row enter the run number `84884` and the
angle `0.8`. Click the process button below the table widget to run the
reduction
9) Check the out workspaces in the ADS. You should see intermediate
workspaces appended in `_lambda`, `_lambda_subtracted)bg` etc.