Add DetermineSpinStateOrder algorithm for PolSans epic (#39600)
Add algorithm to implement SANS scientists' script to automatically
determine the order of spin states for each workspace in a PolSans
transmission run.

The algorithm uses the spin flipper current log to determine the flipper
states. It compares the beam transmission to the average transmission
(over all state combinations) to determine the beam polarisation.