Refactors of `Nexus::File` and co. (#39668)
This takes a lot of refactors that were in PR #39658 , which were
unrelated to the core behavior of that PR.

These refactors are individually small, but it may seem like a lot of
changes.

- create `NexusAddress`, which creates a path-like object to specifying
locations inside a Nexus file
- `NexusFile` uses a `NexusAddress` to hold the current address in the
file
- Cleaned up or removed structures in Nexus layer
- From `napi` removed: `NXgetnextentry`, `NXsameID`, `NXinitgroupdir`,
`NXinitattrdir`, `NXmalloc64`, and `NXfree`. These were either unused,
or used only in a test of`napi`, and so could be safely removed.
- From `napi5` removed: `NXgetnextentry`, `attr_info`, and
`nxgroup_info`. The first was only used in a single test of `napi`,
which could be deleted.
- From `napi_helper` removed: all of the methods listed under
implementation of `NXopenaddress`, used to manipulate addresses or the
group ID stack
- Inside many unit tests, added a `std::flush` command to print-outs.
This makes debugging easier, especially when an error causes a segfault.
- Refactored any place in the code that opened an `H5File` object to
make sure it used `H5Util::defaultFileAcc()` instead of constructing its
own access list, to ensure consistency across the code base.
- Many code quality refactors inside `NexusFile`


Refs #38332 
EWM 11399

### To test:

This is a refactor, so ensure existing tests pass. Also check logic of
new tests.

Please check the refactors inside `NexusFile`, as these are the most
important and substantial refactors.

This does not require release notes because it is a purely internal
issues