Avoid Linux standalone launching error (#39670)
On some Linux distributions (reported by users on Ubuntu 22.04 and
AlmaLinux 9.6), there are some files missing from the `share` folder,
which prevents mantidworkbench from launching with the following error:
`xkbcommon: ERROR: failed to add default include path
/jenkins_workdir/workspace/release-next_nightly_deployment/_bundle_build/mantidworkbench/share/X11/xkb`
In the packaging script that generates the standalone installer, we
currently trim down the share folder to cut down on the size of the
package.

The easiest and most maintainable solution is to keep all of the share
directory when creating standalone installers. There have been two
releases now where we have had to add extra things that have been
missing because we trim the share folder. We only save 20mb by trimming
it down, and the risk is too high to justify that.