Remove link between muon and curvefitting libs (#39656)
This PR removes the use of `Mantid::CurveFitting::EigenMatrix` from the
`Muon` library to avoid the linking of two plugin libraries.

We use `PartialPivLU` decomposition in calculating the inverse,
replicating what the `Mantid::CurveFitting::EigenMatrix` is doing under
the hood.

`long double` is used on arm mac, as it brings precision of eigen matrix
multiplication inline (lowers it) with the other operating systems. I'm
not sure why, you would expect a greater degree of precision.

Here is a compiler explorer used to interrogate the differences when
different types are used with eigen across our compilers:
https://godbolt.org/z/51W6KPecb

This illustrates the issue, but doesn't show the same impact on arm64 of
switching to `long double`.