Remove `napi5` and most of `napi` test layer (#39691)
### Description of work

#### Summary of work

With PR #39657, now `napi5` is almost completely unused, and `napi` is
only being used for `NXopenaddress` (to be removed in PR #39658). It is
therefore safe at this point to remove `napi5` and most of the tests of
`napi`.

The following have been left in the `napi` layer:
- `napi` because `NXopenaddress` is still being called, and it is needed
for the remaining tests to work
- `NapiUnitTest` because it may have useful tests to be duplicated in
`NexusFileTest`, and contains some notes on proper functionality
- `napi_test_cpp` because it is actually a test of `Nexus::File`
- `test_nxunlimited` because it has a useful test of unlimited
dimensions that doesn't seem to be replicated in other tests

Part of #38332
EWM 11399

#### Further detail of work

Lots of deletions.  It feels good.

Note that `napi.cpp` now implements `NXgetdata` and `NXgetchardata` from
`napi5`. These methods now *only* show up in the tests of the `napi`
layer, so it doesn't particularly matter how this is implemented at this
point so long as the tests pass.

### To test:

This is a refactor. The existing tests passing, in particular tests in
the Nexus and DataHandling layers, should suffice.

This does not require release notes because it is a purely internal
matter.
<!-- -->

---

### Reviewer

Please comment on the points listed below ([full
description](http://developer.mantidproject.org/ReviewingAPullRequest.html)).
**Your comments will be used as part of the gatekeeper process, so
please comment clearly on what you have checked during your review.** If
changes are made to the PR during the review process then your final
comment will be the most important for gatekeepers. In this comment you
should make it clear why any earlier review is still valid, or confirm
that all requested changes have been addressed.

#### Code Review

- Is the code of an acceptable quality?
- Does the code conform to the [coding
standards](http://developer.mantidproject.org/Standards/)?
- Are the unit tests small and test the class in isolation?
- If there is GUI work does it follow the [GUI
standards](http://developer.mantidproject.org/Standards/GUIStandards.html)?
- If there are changes in the release notes then do they describe the
changes appropriately?
- Do the release notes conform to the [release notes
guide](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)?

#### Functional Tests

- Do changes function as described? Add comments below that describe the
tests performed?
- Do the changes handle unexpected situations, e.g. bad input?
- Has the relevant (user and developer) documentation been
added/updated?

Does everything look good? Mark the review as **Approve**. A member of
`@mantidproject/gatekeepers` will take care of it.

### Gatekeeper

If you need to request changes to a PR then please add a comment and set
the review status to "Request changes". This will stop the PR from
showing up in the list for other gatekeepers.