Add unit test for linearAbsorpCoef function (#39703)
When testing the new windows CI image running with docker 17 we fell
foul of a compiler bug in a recent version of MSVC:

https://developercommunity.visualstudio.com/t/Serious-compiler-bug-in-144435207-duri/10908277?sort=newest&viewtype=solutions

This bug caused errors in tests that touch upon the
`AbsorptionCorrection` class, which themselves are several steps removed
from the root cause of the issue.

We will look to avoid the compiler verison with the bug, but to make
this error easier to identify in the future, I've added a unit test
which tests the offending function directly.