`Nexus::File` swallows `NexusFile5` (#39715)
`napi` formerly worked using a `struct` called `NexusFile5` which
contained information about the currently opened locations and current
address. This structure lived as a `shared_ptr` inside `Nexus::File`,
and was passed as an argument to the `napi` functions.

Keeping it as a separate `struct` with a `shared_ptr` can cause
conflicts when an algorithm has two instances of the same open file.
This occurs in the algorithms that use both `NexusClasses` and
`Nexus::File`, which can alter the shared struct in an inconsistent
manner. Using a `unique_ptr` creates strange and [un-debuggable errors
in
Windows](https://builds.mantidproject.org/job/pull_requests-conda-windows/11496/testReport/junit/projectroot.scripts/test/python_scripts_SANSUtilityTest_SANSUtilityTest/).

Previously, we were prevented from moving the variables of `NexusFile5`
into `Nexus::File` because of the constraint of needing to interface
withe `napi` through the struct.

This PR removes almost the entirety of `napi`. Those parts which remain
have been invertedto call `Nexus::File` instead.

Part of #38332
EWM 11399

#### Further detail of work

This retains a few `napi` functions which are only called in tests, but
inverts them to call `Nexus::File`. This is a halfway measure.

The `napi` tests retained are those in `NapiUnitTest` related
specifically to string data and attributes. These tests presented the
main difficulties during refactoring, and have some notes about how the
program used to function. It might be worthwhile to retain these
portions of the code, at least until manual testing of 6.14.

### To test:

This is a refactor.  Ensure all current tests pass.

This does not require release notes because this is an internal issue
that does not impact users.