Fix msvc errors seen on updated windows jenkins node (#39624)
This PR resolves several compilation warnings seen in the latest version
of the MSVC compiler.

If visual studio is updated, these warnings are visible on local
developer builds.

This work is required to build mantid on the new java17 windows docker
images, as they download an up to date version of MSVC.

Enables https://github.com/mantidproject/dockerfiles/issues/109

Co-authored-by: sf1919 <sarah.foxley@stfc.ac.uk>