Move classes outside of AlignAndFocusPowderSlim into own folder. (#39701)
### Description of work

~~This includes the changes in #39684 and should be reviewed after
that.~~

This is similar to how `LoadRaw` is done.

This move the classes `NexusLoader`, `BankCalibration`,
`ProcessEventsTask` and `ProcessBankTask` out of the
`AlignAndFocusPowderSlim` algorithm file and into their own files in a
sub-directory call `AlignAndFocusPowderSlim`. This also moves
everything, including the algorithm itself, into the
`AlignAndFocusPowderSlim` namespace under `DataHandling`.

#### Summary of work

*There is no associated issue.*


#### Further detail of work

### To test:

Code review

*This does not require release notes* because internal change only