Add AssertSpinStateOrder algorithm (#39654)
Add algorithm to assert the order of Polarised SANS transmission run
periods via their spin states. Uses the new DetermineSpinStateOrder
(#39600) algorithm.

If the periods are in the incorrect order and the `Reorder` parameter is
set to `True`, then the child workspaces are reordered. To do this, I
have added a method to `WorkspaceGroup`.

Co-authored-by: Adri Diaz <146007827+adriazalvarez@users.noreply.github.com>