`Nexus::File` to use a shared pointer to a `FileID` object (#39732)
### Description of work

#### Summary of work

Currently `Nexus::File` uses an `hid_t` as an identifier to an open HDF5
file object. However, this could potentially cause problems with the
close functionality, leading to a memory leak of no one closes the file,
or leading to someone closing the file while other tasks are using it.

Using an `H5File` object will ensure that the file is closed (as opposed
to the `hid_t` deleted) when the last owner to the pointer releases it.
However, `H5File` is part of the C++ library, and contains a lot of
other memory overhead that is not strictly needed in this instance.

Instead a new class `FileID` was created which does exactly what is
needed, which is to call `H5Fclose()` when it deletes. Now `Nexus::File`
has a `shared_ptr<FileID>` which can easily handle sharing the same file
while making sure the last user closes it.

While doing this, I realized the `Nexus::File` assignment operator can
be deleted. It is not used, and using it could create strange conditions
for the file ID.

#### Purpose of work

This is part of #38332 
EWM 11399

#### Further detail of work

The variable `m_fid`, along with `m_groupaddr` will be deleted in the
next PR in this series.

There were not formerly unit tests of the rule of three. This adds tests
of copying and deletion of `Nexus::File`

### To test:

This is a refactor.  Ensure all unit and system tests pass.

This does not require release notes because it is an internal matter.

---

### Reviewer

Please comment on the points listed below ([full
description](http://developer.mantidproject.org/ReviewingAPullRequest.html)).
**Your comments will be used as part of the gatekeeper process, so
please comment clearly on what you have checked during your review.** If
changes are made to the PR during the review process then your final
comment will be the most important for gatekeepers. In this comment you
should make it clear why any earlier review is still valid, or confirm
that all requested changes have been addressed.

#### Code Review

- Is the code of an acceptable quality?
- Does the code conform to the [coding
standards](http://developer.mantidproject.org/Standards/)?
- Are the unit tests small and test the class in isolation?
- If there is GUI work does it follow the [GUI
standards](http://developer.mantidproject.org/Standards/GUIStandards.html)?
- If there are changes in the release notes then do they describe the
changes appropriately?
- Do the release notes conform to the [release notes
guide](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)?

#### Functional Tests

- Do changes function as described? Add comments below that describe the
tests performed?
- Do the changes handle unexpected situations, e.g. bad input?
- Has the relevant (user and developer) documentation been
added/updated?

Does everything look good? Mark the review as **Approve**. A member of
`@mantidproject/gatekeepers` will take care of it.

### Gatekeeper

If you need to request changes to a PR then please add a comment and set
the review status to "Request changes". This will stop the PR from
showing up in the list for other gatekeepers.