Restore reverted change: Abins refactor replacing SData classes with Euphonic features (#39669)
### Description of work

This reinstates #37350 , which was reverted in #39665

This was done in order to stop breakage of package build tests on
non-Linux platforms.
It seems this is caused by an incompatibility between multiprocessing
"spawn" and the systemtest code, which also spawns Python subprocesses.
This mechanism is not supposed to be nested. On Linux the default
mechanism is "fork" which seems to work in parallel systemtest, but from
Python 3.14 this default will also change to "spawn" and start failing.

#### Summary of work
- The branch from #37350 has been rebased back onto `main`
- The conditions for hanging test suites were investigated
- Multiprocessing code is removed from Abins and replaced by serial
`map` operations.
- The performance impact seems to be small, due to optimisations in
Euphonic which came after the parallel operations were originally
introduced to the branch.

#### Purpose of work
See #37350

#### Further detail of work
It's not great that multiprocessing is incompatible with the systemtest
suite. In this case we can manage without, but there is a good chance we
will want to use it in future. Multithreading is probably ok, but these
are not equivalent tools.

### To test:

Everything is covered by CI. In addition to the default stuff we need to
check `build_packages_from_branch` on non-Linux platforms as this was
breaking.

This shouldn't require a release note because #37350 didn't.

---

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>