Goodnight, `napi` :last_quarter_moon_with_face:  (#39658)
### Description of work

#### Summary of work

Removes `File::getAddress()`, which is the last little bit of `napi`.

What is left is a series of shims to enable three tests leftover in
`NapiUnitTest`

Some files were renamed:
- `napi_helper` --> `hdf5_type_helper` to reflect the fact that the only
functions remaining are for translating between Nexus types, HDF5 types,
or turning generic HDF5 types into the ones we use.
- `napi` --> `inverted_napi` to reflect that these are shims calling
`Nexus::File`, which is the inverse of how they formerly worked

Refs #38332 
EWM 11399

#### Further detail of work
<!-- Please provide a more detailed description of the work that has
been undertaken.
-->

There was a very difficult bug occurring when running `SANSadd2Test`. It
would fail, saying it was unable to open a file due to close degree. The
full stack trace was not printed, but the error was raised in the python
script `file_information.py` in the function
`get_geometry_information_isis_nexus()`, in particular when starting the
context manager `h5.File()`. This opens with default instead of
close-strong access, which creates contradictions that generate an
error.

That is when the error is raised, but the errant condition can be traced
back to the `Load` algorithm, in the method `Load::loadSingleFile()`.

### To test:

Double check the logic of the new `openAddress()` method.

This is a refactor, so all tests should pass.

This does not require release notes because it is a purely internal
matter.

---

### Reviewer

Please comment on the points listed below ([full
description](http://developer.mantidproject.org/ReviewingAPullRequest.html)).
**Your comments will be used as part of the gatekeeper process, so
please comment clearly on what you have checked during your review.** If
changes are made to the PR during the review process then your final
comment will be the most important for gatekeepers. In this comment you
should make it clear why any earlier review is still valid, or confirm
that all requested changes have been addressed.

#### Code Review

- Is the code of an acceptable quality?
- Does the code conform to the [coding
standards](http://developer.mantidproject.org/Standards/)?
- Are the unit tests small and test the class in isolation?
- If there is GUI work does it follow the [GUI
standards](http://developer.mantidproject.org/Standards/GUIStandards.html)?
- If there are changes in the release notes then do they describe the
changes appropriately?
- Do the release notes conform to the [release notes
guide](https://developer.mantidproject.org/Standards/ReleaseNotesGuide.html)?

#### Functional Tests

- Do changes function as described? Add comments below that describe the
tests performed?
- Do the changes handle unexpected situations, e.g. bad input?
- Has the relevant (user and developer) documentation been
added/updated?

Does everything look good? Mark the review as **Approve**. A member of
`@mantidproject/gatekeepers` will take care of it.

### Gatekeeper

If you need to request changes to a PR then please add a comment and set
the review status to "Request changes". This will stop the PR from
showing up in the list for other gatekeepers.