Clean x gap correction
- Set default period to 2 pi
- Clean part to find biggest gap in x range by using numpy detectors

The only change in functunality is that if there are two or more gaps
with the exact same size, the gap chosen might or might not be the same
as before. I don't believe this is a problem, since this is not common
and the point of the routine is to avoid the biggest gap in the x range.
If two such big gaps are present, choosing one over the other should not
improve the visualisation of the detectors one way or the other.