CI Changes to enable use of updated windows runners (#39738)
The following changes have been made as part of the work to update the
windows runners:

- Upgrade the dev cmake preset to match v143 toolset
- Use ninja as a generator on windows in our CI
- Use `target_precompile_header` functionality native to cmake, remove
our custom function.
- Enable precompiled headers (pch) on linux/osx.
- Mark precompiled headers as system headers on linux/osx.

Regarding that last change, the impact is that warnings will be
supressed for all files compiled as part of a pch. This is necessary as
gcc picks up on issues in external code if included in pch, raising
warnings that we cannot solve.
The effect of this, is that it has the potential to disable warnings in
some of our mantid headers that fulfil the following criteria:
- Are defined in header files, included in target X.
- Are include in a PCH.
- That PCH is then set on the target X
- The mantid header is then not recompiled as it is in the PCH, and any
warnings hidden.

It is of note though that any source files (so the majority of the
implementation, where you'd expect the warnings to be) would still raise
warnings. Given this, I think that increased compile speeds may be worth
this.

If we do not agree with this, we can simply redisable pch for linux and
osx in a future PR.

The PR has been tested in pieces due to manipulation of the jenkins tags
to target a specific updated runner:
Windows package build:
https://builds.mantidproject.org/job/build_packages_from_branch/1253/
Unix package builds:
https://builds.mantidproject.org/job/build_packages_from_branch/1256/

Related to https://github.com/mantidproject/dockerfiles/issues/109