Replace NXIformatNeXusTime in NexusFile (#39762)
### Description of work

Every Nexus file is created with an attribute `file_time` that gives the
current local time in the format `YYYY-MM-DDTHH:MM:SS±HH:MM`. This
creates a helper function in `DateAndTime` that will create the
formatted string. This may have benefits elsewhere, the possibilities
are endless.

Relates to [EWM
12326](https://ornlrse.clm.ibmcloud.com/ccm/web/projects/Neutron%20Data%20Project%20%28Change%20Management%29#action=com.ibm.team.workitem.viewWorkItem&id=12326)

```text
This function `NXIformatNeXusTime()` appears in `napi.cpp`:
Its purpose is to return a formatted time string indicating when the file was created.  
It is almost certainly not the best or most modern way to get a string with the current time.
Investigate how it can be rewritten, then add a static function in `NexusFile.cpp` 
to write out the formatted time stamp.
```

### To test:
A test has been created in `DateAndTimeTest`, there is currently no easy
way to perform a manual test.

*This does not require release notes* because **this is an internal
change and the functionality is not exposed**