Nexus to use unsigned dimensions (#39754)
### Description of work

Writing data to NeXus files requires specifying the dimensions of the
data through an array object. `NexusFile_fwd.h` defines a helpful
typedef, `DimVector`, which is a `std::vector`. To be compatible with
`napi` this had been declared as a vector of `int64_t`; however,
dimensions cannot be negative, so this should be a vector of unsigned
ints. Further, the values in this array almost always come from the size
of some other container object, with type `std::size_t`, requiring
unnecessary static casts.

From the start of the Nexus refactor work, it has been an ambition to
change these values to unsigned ints, but has been prohibitively
difficult due to the then condition of `napi`.

This finally gets in that bit of cleanup.

Part of #38332

### To test:

Changes are very mechanical.

Ensure all existing tests can pass without regressions.

This does not need release notes because this is entirely internal