Add validator polsans transmission workspaces (#39770)
Add `PolSANSWorkspaceValidator` to reduce the duplicated validation code
across multiple polsans algorithms.

Closes #39621
### To test:

Test the conditions listed in the original issue by making test
workspaces and giving them to `DetermineSpinStateOrder` (for an example
with a default `PolSANSWorkspaceValidator`) via a python script and the
algorithm gui.
Also test the python algorithm `AssertSpinStateOrder` to check the
python exposing has worked properly.

Code needs careful reviewing to check that the removed code is
equivalent to using the new validator.