Set up QThread using QObject instead of inheriting
Instead of creating a QThread by inheriting QThread (runs on the main
thread), create a QObject that runns on a QThread, which ensures that
worker gets run in a separate thread.

From the C++ side, need to release the GIL so that the python
interpreter can run that section in parallel rather than switching back
and forth between main and separate thread (default behaviour in
python).