Fix jenkinsfile variable scopes to avoid warnings (#39827)
Since the recent Jenkins update, we are presented with warnings of the
form:
```
Did you forget the `def` keyword? WorkflowScript seems to be setting a field named ANACONDA_CHANNEL_DEFAULT (to a value of type String) which could lead to memory leaks or other issues.
```
This is resolved by ensuring that all local variables are scoped
correctly by prepending the `def` keyword where possible. Variables that
are required within helper functions must be prepended with `@Field
<Type>` to make them visible at the class level.

The parameter defaults have been refactored into a map. This map is only
declared in the local scope, which means it is accessible to the
pipeline block (where the parameters are created), but not in any of the
helper functions later in the file, where they are not needed.

Additionally, as per the Jenkins
[documentation](https://www.jenkins.io/doc/book/pipeline/syntax/#parameters),
the `params` object is used instead of the 'env' keyword when
referencing pipeline parameters. This has the advantage of allowing us
to make use of the type of the parameter, instead of everything being
converted to a string in the case of using `env`.